+++ tests/integration/_expected_outputs/sourcemaps/sourcemaps-inject-malformed-map/app.js
@@ -1,0 +1,6 @@
+
+!function(){try{var e="undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof globalThis?globalThis:"undefined"!=typeof self?self:{},n=(new e.Error).stack;n&&(e._sentryDebugIds=e._sentryDebugIds||{},e._sentryDebugIds[n]="b5859ee8-d3d3-52fa-9ae8-e1570e2e734f")}catch(e){}}();
+console.log("Hello World");
+//# sourceMappingURL=app.js.map
+
+//# debugId=b5859ee8-d3d3-52fa-9ae8-e1570e2e734f